Can't believe that I am having this problem in so many different places now.

+++ renameExperimentalData.m
@@ -80,8 +80,8 @@ function renameExperimentalData()
                     % temprow = solutionstatestable.getRowAtIndex(i).getAsMat();
                     temprow = temptable.getRowAtIndex(i).getAsMat();
                     temprow2 = org.opensim.modeling.Vector().createFromMat(temprow);
-%                     sto.append(statetime.get(i).doubleValue(), temprow2);
-                    sto.append(statetime.get(i), temprow2);
+                    sto.append(statetime.get(i).doubleValue(), temprow2);
+%                     sto.append(statetime.get(i), temprow2);
 
                 end
